# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Health First API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  wilaya?: string | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  wilaya?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface DashboardSummary {
  medicationsToday: number;
  medicationsTakenToday: number;
  upcomingAppointments: number;
  totalRecords: number;
  familyMembers: number;
  activeMedications: number;
}

export interface Medication {
  id: number;
  userId: number;
  /** @nullable */
  familyMemberId?: number | null;
  name: string;
  dosage: string;
  frequency: string;
  times: string;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  notes?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface CreateMedicationBody {
  name: string;
  dosage: string;
  frequency: string;
  times: string;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  familyMemberId?: number | null;
}

export interface UpdateMedicationBody {
  name?: string;
  dosage?: string;
  frequency?: string;
  times?: string;
  startDate?: string;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  notes?: string | null;
  isActive?: boolean;
}

export interface MedicationReminder {
  id: number;
  medicationId: number;
  medicationName: string;
  dosage: string;
  scheduledTime: string;
  taken: boolean;
  /** @nullable */
  takenAt?: string | null;
  date: string;
  /** @nullable */
  familyMemberName?: string | null;
}

export interface MedicalRecord {
  id: number;
  userId: number;
  /** @nullable */
  familyMemberId?: number | null;
  title: string;
  category: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  fileName?: string | null;
  recordDate: string;
  createdAt: string;
}

export interface CreateRecordBody {
  title: string;
  category: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  fileName?: string | null;
  recordDate: string;
  /** @nullable */
  familyMemberId?: number | null;
}

export interface Doctor {
  id: number;
  name: string;
  specialty: string;
  wilaya: string;
  address: string;
  /** @nullable */
  phone?: string | null;
  rating: number;
  reviewCount: number;
  availableDays: string;
  consultationFee: number;
  /** @nullable */
  imageUrl?: string | null;
  isOnlineConsultation: boolean;
}

export interface Appointment {
  id: number;
  userId: number;
  doctorId: number;
  doctorName: string;
  doctorSpecialty: string;
  appointmentDate: string;
  appointmentTime: string;
  status: string;
  /** @nullable */
  notes?: string | null;
  isOnline: boolean;
  createdAt: string;
}

export interface CreateAppointmentBody {
  doctorId: number;
  appointmentDate: string;
  appointmentTime: string;
  /** @nullable */
  notes?: string | null;
  isOnline?: boolean;
}

export interface UpdateAppointmentBody {
  status?: string;
  /** @nullable */
  notes?: string | null;
  appointmentDate?: string;
  appointmentTime?: string;
}

export interface Medicine {
  name: string;
  available: boolean;
  /** @nullable */
  price?: number | null;
}

export interface Pharmacy {
  id: number;
  name: string;
  wilaya: string;
  address: string;
  /** @nullable */
  phone?: string | null;
  isOpenNow: boolean;
  is24h: boolean;
  medicines: Medicine[];
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
}

export interface FamilyMember {
  id: number;
  userId: number;
  name: string;
  relationship: string;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  bloodType?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface CreateFamilyMemberBody {
  name: string;
  relationship: string;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  bloodType?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateFamilyMemberBody {
  name?: string;
  relationship?: string;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  bloodType?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type ListDoctorsParams = {
  specialty?: string;
  wilaya?: string;
};

export type ListPharmaciesParams = {
  medicine?: string;
  wilaya?: string;
};