# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: My Health First API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /doctors:
    get:
      operationId: listDoctors
      tags: [doctors]
      summary: List all doctors
      parameters:
        - name: specialty
          in: query
          schema:
            type: string
        - name: wilaya
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of doctors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"
  /doctors/{id}:
    get:
      operationId: getDoctor
      tags: [doctors]
      summary: Get a doctor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Doctor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
        "404":
          description: Not found
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments for current user
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Appointment booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
...
  /appointments/{id}:
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Cancel an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cancelled
...
  /family:
    get:
      operationId: listFamilyMembers
      tags: [family]
      summary: List family members
      responses:
        "200":
          description: List of family members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FamilyMember"
    post:
      operationId: createFamilyMember
      tags: [family]
      summary: Add a family member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFamilyMemberBody"
      responses:
        "201":
          description: Family member created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyMember"
  /family/{id}:
    patch:
      operationId: updateFamilyMember
      tags: [family]
      summary: Update a family member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFamilyMemberBody"
      responses:
        "200":
          description: Updated family member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyMember"
        "404":
          description: Not found
    delete:
      operationId: deleteFamilyMember
      tags: [family]
      summary: Remove a family member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: ["string", "null"]
        wilaya:
          type: ["string", "null"]
      required:
        - name
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        wilaya:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - createdAt
...
    Doctor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        specialty:
          type: string
        wilaya:
          type: string
        address:
          type: string
        phone:
          type: ["string", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        availableDays:
          type: string
        consultationFee:
          type: number
        imageUrl:
          type: ["string", "null"]
        isOnlineConsultation:
          type: boolean
      required:
        - id
        - name
        - specialty
        - wilaya
        - address
        - rating
        - reviewCount
        - availableDays
        - consultationFee
        - isOnlineConsultation
    Appointment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        doctorId:
          type: integer
        doctorName:
          type: string
        doctorSpecialty:
          type: string
        appointmentDate:
          type: string
        appointmentTime:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        isOnline:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - userId
        - doctorId
        - doctorName
        - doctorSpecialty
        - appointmentDate
        - appointmentTime
        - status
        - isOnline
        - createdAt
    CreateAppointmentBody:
      type: object
      properties:
        doctorId:
          type: integer
        appointmentDate:
          type: string
        appointmentTime:
          type: string
        notes:
          type: ["string", "null"]
        isOnline:
          type: boolean
      required:
        - doctorId
        - appointmentDate
        - appointmentTime
    UpdateAppointmentBody:
      type: object
      properties:
        status:
          type: string
        notes:
          type: ["string", "null"]
        appointmentDate:
          type: string
        appointmentTime:
          type: string
...
    FamilyMember:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        relationship:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        bloodType:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - name
        - relationship
        - createdAt
    CreateFamilyMemberBody:
      type: object
      properties:
        name:
          type: string
        relationship:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        bloodType:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - name
        - relationship
    UpdateFamilyMemberBody:
      type: object
      properties:
        name:
          type: string
        relationship:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        bloodType:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]